refactor: widen Completion callback result from Int32 to Int
Int is pointer-sized (64-bit on all supported platforms), matching
the kernel's ssize_t return. Removes 5 truncation casts in the
epoll driver and eliminates a theoretical >2GB transfer bug. The
io_uring driver widens the natively-32-bit CQE.res to Int.

+++ boucle/net/probe.mojo
@@ -92,7 +92,7 @@ struct ProbeResult(Movable):
         self.status = move.status
 
 
-def result_from_connect_cqe(result: Int32) -> PortStatus:
+def result_from_connect_cqe(result: Int) -> PortStatus:
     """Map a connect(2) CQE result code to a PortStatus.
 
     Interprets the kernel return value from an async connect operation:
@@ -106,9 +106,9 @@ def result_from_connect_cqe(result: Int32) -> PortStatus:
     Returns:
         The corresponding PortStatus.
     """
-    if result == Int32(0):
+    if result == 0:
         return PortStatus.OPEN
-    if result == Int32(-111):
+    if result == -111:
         return PortStatus.CLOSED
     return PortStatus.FILTERED
