+++ boucle/watch/recv.mojo
@@ -35,7 +35,7 @@ struct _RecvFutureState(_FutureCallback):
     """
 
     var completion: Completion
-    var _cqe_result: Int32
+    var _cqe_result: Int
     var done: Bool
     var consumed: Bool
 
@@ -46,7 +46,7 @@ struct _RecvFutureState(_FutureCallback):
         must wire invoke and context after heap allocation.
         """
         self.completion = Completion()
-        self._cqe_result = Int32(0)
+        self._cqe_result = 0
         self.done = False
         self.consumed = False
 
@@ -61,7 +61,7 @@ struct _RecvFutureState(_FutureCallback):
         self.done = move.done
         self.consumed = move.consumed
 
-    def set_result(mut self, result: Int32):
+    def set_result(mut self, result: Int):
         """Store the raw CQE result from io_uring recv.
 
         Args:
