+++ tests/boucle/proactor/test_completion.mojo
@@ -8,20 +8,20 @@ from boucle.proactor.completion import Completion
 struct Tracker:
     """Records callback invocations for test assertions."""
 
-    var last_result: Int32
+    var last_result: Int
     var last_flags: UInt32
     var count: Int
 
     def __init__(out self):
         """Construct a zeroed tracker."""
-        self.last_result = Int32(0)
+        self.last_result = 0
         self.last_flags = UInt32(0)
         self.count = 0
 
     @staticmethod
     def on_complete(
         ctx: Pointer[NoneType, MutUntrackedOrigin],
-        result: Int32,
+        result: Int,
         flags: UInt32,
     ):
         """Callback that records result into the Tracker."""
@@ -42,14 +42,14 @@ def test_completion() raises:
     var cmp = Completion(invoke=Tracker.on_complete, context=ctx)
 
     # Fire with known values.
-    cmp.fire(result=Int32(42), flags=UInt32(7))
+    cmp.fire(result=42, flags=UInt32(7))
 
     assert_equal(tracker.count, 1)
     assert_equal(Int(tracker.last_result), 42)
     assert_equal(Int(tracker.last_flags), 7)
 
     # Fire again with different values.
-    cmp.fire(result=Int32(-111), flags=UInt32(0))
+    cmp.fire(result=-111, flags=UInt32(0))
     assert_equal(tracker.count, 2)
     assert_equal(Int(tracker.last_result), -111)
